refactor: Split CLIcore monolithic sources
Extract CLIcore_script intercept logic, CLIcore_UI command execution,
and cli_calc_parser mathematical functions and evaluation logic into
smaller, focused modules.
