+++ src/CLImain.c
@@ -17,7 +17,7 @@
 
 #include "CLIcore.h"
 
-#include "CLIcore/CLIcore_UI.h"
+#include "CLIcore/CLIcore_UI_execute.h"
 
 #define STYLE_BOLD    "\033[1m"
 #define STYLE_NO_BOLD "\033[22m"
