+++ src/cli/CLIcore/CLIcore/CLIcore_UI_builtins.c
@@ -9,7 +9,7 @@
 #include <readline/readline.h>
 #endif
 #include "CLIcore.h"
-#include "CLIcore_UI.h"
+#include "CLIcore_UI_execute.h"
 #include "CLIcore_script.h"
 #include "CLIcore/cli_calc_parser.h"
 #include <glob.h>
