+++ src/cli/CLIcore/CLIcore/CLIcore_UI_execute.c
@@ -21,7 +21,7 @@
 #include "CLIcore.h"
 #include "CLIcore/cli_calc_parser.h"
 #include "CLIcore_script.h"
-#include "CLIcore_UI.h"
+#include "CLIcore_UI_execute.h"
 
 #include <fnmatch.h>
 #include <glob.h>
