+++ src/cli/CLIcore/CLIcore/CLIcore_UI_highlight.c
@@ -30,7 +30,7 @@
 #endif
 
 #include "CLIcore.h"
-#include "CLIcore_UI.h"
+#include "CLIcore_UI_execute.h"
 
 
 /**
