+++ src/cli/CLIcore/CLIcore/CLIcore_UI_hintarea.c
@@ -44,7 +44,7 @@
 #endif
 
 #include "CLIcore.h"
-#include "CLIcore_UI.h"
+#include "CLIcore_UI_execute.h"
 
 
 #ifdef USE_READLINE
