+++ src/cli/CLIcore/CLIcore/CLIcore_UI_history.c
@@ -14,7 +14,7 @@
 #include <readline/readline.h>
 #endif
 #include "CLIcore.h"
-#include "CLIcore_UI.h"
+#include "CLIcore_UI_execute.h"
 #include <glob.h>
 #include <sys/wait.h>
 #include "timeutils.h"
