+++ src/cli/CLIcore/CLIcore/CLIcore_script_builtin.c
@@ -4,7 +4,7 @@
 #include <sys/types.h>
 #include <unistd.h>
 #include "CLIcore.h"
-#include "CLIcore_UI.h"
+#include "CLIcore_UI_execute.h"
 #include "CLIcore_script.h"
 #include "CLIcore/cli_calc_parser.h"
