+++ src/cli/CLIcore/CLIcore/CLIcore_script_var.c
@@ -6,7 +6,7 @@
 #include <ctype.h>
 #include <wordexp.h>
 #include "CLIcore.h"
-#include "CLIcore_UI.h"
+#include "CLIcore_UI_execute.h"
 #include "CLIcore_script.h"
 #include "CLIcore/cli_calc_parser.h"
