+++ src/cli/CLIcore/CLIcore/CLIcore_signals.c
@@ -12,7 +12,7 @@
 
 #include "CLIcore.h"
 
-#include "CLIcore_UI.h"
+#include "CLIcore_UI_execute.h"
 
 #include "timeutils.h"
