+++ src/cli/CLIcore/CLIcore/script_includes.h
@@ -4,6 +4,6 @@
 #include <sys/types.h>
 #include <unistd.h>
 #include "CLIcore.h"
-#include "CLIcore_UI.h"
+#include "CLIcore_UI_execute.h"
 #include "CLIcore_script.h"
 #include "CLIcore/cli_calc_parser.h"
