+++ src/cli/CLIcore/cli_calc_parser.c
@@ -28,7 +28,7 @@
 #include "cli_calc_tokenizer.h"
 #include "cli_calc_parser.h"
 #include "CLIcore_script.h"
-
+#include "cli_calc_internal.h"
 /* --------------------------------------------------------
  * Parser state
  * -------------------------------------------------------- */
@@ -37,59 +37,30 @@
 static cli_token  parse_tokens[CLI_CALC_MAX_TOKENS];
 static int        parse_pos;
 static int        parse_ntok;
-static int        parse_error;
+int               parse_error;
 
 /** token stream and current index for cli_calc_eval_line */
 static cli_token  eval_tokens[CLI_CALC_MAX_TOKENS];
 static int        eval_pos;
 static int        eval_ntok;
-static int        eval_error;
+int               eval_error;
 
 
 static char calctmpimname[200];
 
 /* --------------------------------------------------------
- * Value types for the expression evaluator
- * -------------------------------------------------------- */
-
-/** Expression result type tag */
-typedef enum
-{
-    VAL_LONG,
-    VAL_DOUBLE,
-    VAL_STRING,
-    VAL_GENERIC // For functions that don't return a specific value type
-} val_type;
-
-/** Expression result value */
-typedef struct
-{
-    val_type type;
-    long     lval;
-    double   dval;
-    char     sval[CLI_CALC_TOKEN_MAXLEN];
-} val_t;
-
-/* --------------------------------------------------------
- * Forward declarations
- * -------------------------------------------------------- */
-
-static val_t parse_expr(int min_prec);
-static val_t parse_primary(void);
-
-/* --------------------------------------------------------
  * Token stream helpers
  * -------------------------------------------------------- */
 
-static int parse_mode = 0; // 0 for cli_parse, 1 for cli_calc_eval_line
+int parse_mode = 0; // 0 for cli_parse, 1 for cli_calc_eval_line
 
 // Helper functions for cli_parse
-static inline cli_token *cur_parse(void)
+cli_token *cur_parse(void)
 {
     return &parse_tokens[parse_pos];
 }
 
-static inline cli_token *advance_parse(void)
+cli_token *advance_parse(void)
 {
     cli_token *t = &parse_tokens[parse_pos];
     if (parse_pos < parse_ntok)
@@ -100,12 +71,12 @@ static inline cli_token *advance_parse(void)
 }
 
 // Helper functions for cli_calc_eval_line
-static inline cli_token *cur_eval(void)
+cli_token *cur_eval(void)
 {
     return &eval_tokens[eval_pos];
 }
 
-static inline cli_token *advance_eval(void)
+cli_token *advance_eval(void)
 {
     cli_token *t = &eval_tokens[eval_pos];
     if (eval_pos < eval_ntok)
@@ -119,7 +90,7 @@ static inline cli_token *advance_eval(void)
 /**
  * @brief Report a parse error and set error flag
  */
-static void parse_errmsg(const char *msg)
+void parse_errmsg(const char *msg)
 {
     if (parse_mode == 1 || data.core.Debug > 0)
     {
@@ -132,75 +103,6 @@ static void parse_errmsg(const char *msg)
     }
 }
 
-/* --------------------------------------------------------
- * Operator precedence table
- *
- * Higher number = tighter binding.
- * Right-associative operators use prec - 1 for the
- * right-hand parse.
- *
- *   + -    : 1
- *   * /    : 2
- *   ^      : 3 (right-assoc)
- * -------------------------------------------------------- */
-
-static inline int get_prec(cli_token_type t)
-{
-    switch (t)
-    {
-        case TOK_OP_QUESTION:
-            return -1;
-        case TOK_OP_OR:
-            return 1;
-        case TOK_OP_AND:
-            return 2;
-        case TOK_OP_BOR:
-            return 3;
-        case TOK_OP_BXOR:
-            return 4;
-        case TOK_OP_BAND:
-            return 5;
-        case TOK_OP_EQ:
-        case TOK_OP_NEQ:
-            return 6;
-        case TOK_OP_LT:
-        case TOK_OP_LE:
-        case TOK_OP_GT:
-        case TOK_OP_GE:
-            return 7;
-        case TOK_OP_LSHIFT:
-        case TOK_OP_RSHIFT:
-            return 8;
-        case TOK_OP_PLUS:
-        case TOK_OP_MINUS:
-            return 9;
-        case TOK_OP_STAR:
-        case TOK_OP_SLASH:
-        case TOK_OP_MOD:
-            return 10;
-        case TOK_OP_CARET:
-            return 11;
-        case TOK_EQUAL:
-        case TOK_OP_PLUS_EQ:
-        case TOK_OP_MINUS_EQ:
-        case TOK_OP_STAR_EQ:
-        case TOK_OP_SLASH_EQ:
-            return 0;
-        default:
-            return -1;
-    }
-}
-
-static inline int is_right_assoc(cli_token_type t)
-{
-    return (t == TOK_OP_CARET)
-        || (t == TOK_EQUAL)
-        || (t == TOK_OP_PLUS_EQ)
-        || (t == TOK_OP_MINUS_EQ)
-        || (t == TOK_OP_STAR_EQ)
-        || (t == TOK_OP_SLASH_EQ)
-        || (t == TOK_OP_QUESTION);
-}
 
 /* --------------------------------------------------------
  * Type promotion helpers
@@ -213,7 +115,7 @@ static inline int is_right_assoc(cli_token_type t)
 /**
  * @brief Promote a val_t to double if it is long
  */
-static inline double to_double(val_t v)
+double to_double(val_t v)
 {
     if (v.type == VAL_LONG)
     {
@@ -225,7 +127,7 @@ static inline double to_double(val_t v)
 /**
  * @brief Make a long result
  */
-static inline val_t mk_long(long v)
+val_t mk_long(long v)
 {
     val_t r;
     r.type = VAL_LONG;
@@ -238,7 +140,7 @@ static inline val_t mk_long(long v)
 /**
  * @brief Make a double result
  */
-static inline val_t mk_double(double v)
+val_t mk_double(double v)
 {
     val_t r;
     r.type = VAL_DOUBLE;
@@ -251,7 +153,7 @@ static inline val_t mk_double(double v)
 /**
  * @brief Make a string/image result
  */
-static inline val_t mk_string(const char *s)
+val_t mk_string(const char *s)
 {
     val_t r;
     r.type = VAL_STRING;
@@ -269,7 +171,7 @@ static inline val_t mk_string(const char *s)
  * If the image is not found, reports a parse error
  * and returns 0.
  */
-static int check_image(const char *name)
+int check_image(const char *name)
 {
     if (image_ID(
             name,
@@ -288,7 +190,7 @@ static int check_image(const char *name)
 /**
  * @brief Allocate a temporary image name
  */
-static const char *alloc_tmpname(void)
+const char *alloc_tmpname(void)
 {
     snprintf(calctmpimname, 200,
              "_tmpcalc%ld",
@@ -304,1871 +206,6 @@ static const char *alloc_tmpname(void)
  * for +, -, *, /, ^ to match the bison grammar exactly.
  * -------------------------------------------------------- */
 
-/**
- * @brief Evaluate a binary operation
- *
- * Implements type-aware arithmetic matching the
- * original bison grammar's precedence and type rules.
- */
-static val_t eval_binop(
-    cli_token_type op,
-    val_t left,
-    val_t right
-)
-{
-    /* string (image) operands */
-    if (left.type == VAL_STRING
-        || right.type == VAL_STRING)
-    {
-        const char *tmpn = alloc_tmpname();
-
-        /* image OP image */
-        if (left.type == VAL_STRING
-            && right.type == VAL_STRING)
-        {
-            if (!check_image(left.sval)
-                || !check_image(right.sval))
-            {
-                return mk_string("");
-            }
-            switch (op)
-            {
-                case TOK_OP_PLUS:
-                    arith_image_add(
-                        left.sval,
-                        right.sval,
-                        tmpn
-                    );
-                    break;
-                case TOK_OP_MINUS:
-                    arith_image_sub(
-                        left.sval,
-                        right.sval,
-                        tmpn
-                    );
-                    break;
-                case TOK_OP_STAR:
-                    arith_image_mult(
-                        left.sval,
-                        right.sval,
-                        tmpn
-                    );
-                    break;
-                case TOK_OP_SLASH:
-                    arith_image_div(
-                        left.sval,
-                        right.sval,
-                        tmpn
-                    );
-                    break;
-                case TOK_OP_MOD:
-                    arith_image_fmod(left.sval, right.sval, tmpn);
-                    break;
-                case TOK_OP_CARET:
-                    arith_image_pow(left.sval, right.sval, tmpn);
-                    break;
-                case TOK_OP_LT:
-                    arith_image_testlt(left.sval, right.sval, tmpn);
-                    break;
-                case TOK_OP_LE:
-                    arith_image_testle(left.sval, right.sval, tmpn);
-                    break;
-                case TOK_OP_GT:
-                    arith_image_testmt(left.sval, right.sval, tmpn);
-                    break;
-                case TOK_OP_GE:
-                    arith_image_testge(left.sval, right.sval, tmpn);
-                    break;
-                case TOK_OP_EQ:
-                    arith_image_teste(left.sval, right.sval, tmpn);
-                    break;
-                case TOK_OP_NEQ:
-                    arith_image_testne(left.sval, right.sval, tmpn);
-                    break;
-                case TOK_OP_AND:
-                    arith_image_and(left.sval, right.sval, tmpn);
-                    break;
-                case TOK_OP_OR:
-                    arith_image_or(left.sval, right.sval, tmpn);
-                    break;
-                default:
-                    parse_errmsg(
-                        "Unsupported image op"
-                    );
-                    return mk_string("");
-            }
-            return mk_string(tmpn);
-        }
-
-        /* image OP scalar */
-        if (left.type == VAL_STRING)
-        {
-            if (!check_image(left.sval))
-            {
-                return mk_string("");
-            }
-            double rv = to_double(right);
-
-            switch (op)
-            {
-                case TOK_OP_PLUS:
-                    arith_image_cstadd(
-                        left.sval, rv, tmpn);
-                    break;
-                case TOK_OP_MINUS:
-                    arith_image_cstadd(
-                        left.sval, -rv, tmpn);
-                    break;
-                case TOK_OP_STAR:
-                    arith_image_cstmult(
-                        left.sval, rv, tmpn);
-                    break;
-                case TOK_OP_SLASH:
-                    arith_image_cstdiv(
-                        left.sval, rv, tmpn);
-                    break;
-                case TOK_OP_MOD:
-                    arith_image_cstfmod(
-                        left.sval, rv, tmpn);
-                    break;
-                case TOK_OP_CARET:
-                    arith_image_cstpow(
-                        left.sval, rv, tmpn);
-                    break;
-                case TOK_OP_LT:
-                    arith_image_csttestlt(left.sval, rv, tmpn);
-                    break;
-                case TOK_OP_LE:
-                    arith_image_csttestle(left.sval, rv, tmpn);
-                    break;
-                case TOK_OP_GT:
-                    arith_image_csttestmt(left.sval, rv, tmpn);
-                    break;
-                case TOK_OP_GE:
-                    arith_image_csttestge(left.sval, rv, tmpn);
-                    break;
-                case TOK_OP_EQ:
-                    arith_image_cstteste(left.sval, rv, tmpn);
-                    break;
-                case TOK_OP_NEQ:
-                    arith_image_csttestne(left.sval, rv, tmpn);
-                    break;
-                case TOK_OP_AND:
-                    arith_image_cstand(left.sval, rv, tmpn);
-                    break;
-                case TOK_OP_OR:
-                    arith_image_cstor(left.sval, rv, tmpn);
-                    break;
-                default:
-                    parse_errmsg(
-                        "Unsupported image op"
-                    );
-                    return mk_string("");
-            }
-            return mk_string(tmpn);
-        }
-
-        /* scalar OP image */
-        if (right.type == VAL_STRING)
-        {
-            if (!check_image(right.sval))
-            {
-                return mk_string("");
-            }
-            double lv = to_double(left);
-
-            switch (op)
-            {
-                case TOK_OP_PLUS:
-                    arith_image_cstadd(
-                        right.sval, lv, tmpn);
-                    break;
-                case TOK_OP_MINUS:
-                    arith_image_cstsubm(
-                        right.sval, lv, tmpn);
-                    break;
-                case TOK_OP_STAR:
-                    arith_image_cstmult(
-                        right.sval, lv, tmpn);
-                    break;
-                case TOK_OP_SLASH:
-                    arith_image_cstdiv1(
-                        right.sval, lv, tmpn);
-                    break;
-                case TOK_OP_LT:
-                    arith_image_csttestmt(right.sval, lv, tmpn);
-                    break;
-                case TOK_OP_LE:
-                    arith_image_csttestge(right.sval, lv, tmpn);
-                    break;
-                case TOK_OP_GT:
-                    arith_image_csttestlt(right.sval, lv, tmpn);
-                    break;
-                case TOK_OP_GE:
-                    arith_image_csttestle(right.sval, lv, tmpn);
-                    break;
-                case TOK_OP_EQ:
-                    arith_image_cstteste(right.sval, lv, tmpn);
-                    break;
-                case TOK_OP_NEQ:
-                    arith_image_csttestne(right.sval, lv, tmpn);
-                    break;
-                case TOK_OP_AND:
-                    arith_image_cstand(right.sval, lv, tmpn);
-                    break;
-                case TOK_OP_OR:
-                    arith_image_cstor(right.sval, lv, tmpn);
-                    break;
-                default:
-                    parse_errmsg(
-                        "Unsupported image op"
-                    );
-                    return mk_string("");
-            }
-            return mk_string(tmpn);
-        }
-    }
-
-    /* Both numeric (no string) */
-
-    /* long OP long -> long (except / -> double) */
-    if (left.type == VAL_LONG
-        && right.type == VAL_LONG
-        && op != TOK_OP_SLASH)
-    {
-        switch (op)
-        {
-            case TOK_OP_PLUS:
-                return mk_long(
-                    left.lval + right.lval
-                );
-            case TOK_OP_MINUS:
-                return mk_long(
-                    left.lval - right.lval
-                );
-            case TOK_OP_STAR:
-                return mk_long(
-                    left.lval * right.lval
-                );
-            case TOK_OP_MOD:
-                if (right.lval == 0) {
-                    parse_errmsg("Modulo by zero");
-                    return mk_long(0);
-                }
-                return mk_long(
-                    left.lval % right.lval
-                );
-            case TOK_OP_CARET:
-                return mk_long(
-                    (long) pow(
-                        left.lval, right.lval)
-                );
-            case TOK_OP_LT:
-                return mk_long(left.lval < right.lval ? 1 : 0);
-            case TOK_OP_LE:
-                return mk_long(left.lval <= right.lval ? 1 : 0);
-            case TOK_OP_GT:
-                return mk_long(left.lval > right.lval ? 1 : 0);
-            case TOK_OP_GE:
-                return mk_long(left.lval >= right.lval ? 1 : 0);
-            case TOK_OP_EQ:
-                return mk_long(left.lval == right.lval ? 1 : 0);
-            case TOK_OP_NEQ:
-                return mk_long(left.lval != right.lval ? 1 : 0);
-            case TOK_OP_AND:
-                return mk_long((left.lval && right.lval) ? 1 : 0);
-            case TOK_OP_OR:
-                return mk_long((left.lval || right.lval) ? 1 : 0);
-            case TOK_OP_BAND:
-                return mk_long(
-                    left.lval & right.lval);
-            case TOK_OP_BOR:
-                return mk_long(
-                    left.lval | right.lval);
-            case TOK_OP_BXOR:
-                return mk_long(
-                    left.lval ^ right.lval);
-            case TOK_OP_LSHIFT:
-                return mk_long(
-                    left.lval << right.lval);
-            case TOK_OP_RSHIFT:
-                return mk_long(
-                    left.lval >> right.lval);
-            default:
-                break;
-        }
-    }
-
-    /* any numeric mix -> double */
-    {
-        double lv = to_double(left);
-        double rv = to_double(right);
-
-        switch (op)
-        {
-            case TOK_OP_PLUS:
-                return mk_double(lv + rv);
-            case TOK_OP_MINUS:
-                return mk_double(lv - rv);
-            case TOK_OP_STAR:
-                return mk_double(lv * rv);
-            case TOK_OP_SLASH:
-                if (rv == 0.0) {
-                    parse_errmsg("Division by zero");
-                    return mk_double(0);
-                }
-                return mk_double(lv / rv);
-            case TOK_OP_MOD:
-                if (rv == 0.0) {
-                    parse_errmsg("Modulo by zero");
-                    return mk_double(0);
-                }
-                return mk_double(fmod(lv, rv));
-            case TOK_OP_CARET:
-                return mk_double(pow(lv, rv));
-            case TOK_OP_LT:
-                return mk_long(lv < rv ? 1 : 0);
-            case TOK_OP_LE:
-                return mk_long(lv <= rv ? 1 : 0);
-            case TOK_OP_GT:
-                return mk_long(lv > rv ? 1 : 0);
-            case TOK_OP_GE:
-                return mk_long(lv >= rv ? 1 : 0);
-            case TOK_OP_EQ:
-                return mk_long(lv == rv ? 1 : 0);
-            case TOK_OP_NEQ:
-                return mk_long(lv != rv ? 1 : 0);
-            case TOK_OP_AND:
-                return mk_long((lv != 0 && rv != 0) ? 1 : 0);
-            case TOK_OP_OR:
-                return mk_long((lv != 0 || rv != 0) ? 1 : 0);
-            case TOK_OP_BAND:
-                return mk_long(
-                    (long) lv & (long) rv);
-            case TOK_OP_BOR:
-                return mk_long(
-                    (long) lv | (long) rv);
-            case TOK_OP_BXOR:
-                return mk_long(
-                    (long) lv ^ (long) rv);
-            case TOK_OP_LSHIFT:
-                return mk_long(
-                    (long) lv << (long) rv);
-            case TOK_OP_RSHIFT:
-                return mk_long(
-                    (long) lv >> (long) rv);
-            default:
-                break;
-        }
-    }
-
-    parse_errmsg("Unknown operator");
-    return mk_long(0);
-}
-
-/* --------------------------------------------------------
- * Function call parsing
- *
- * Called when a TOK_FUNC_* token has been consumed.
- * Parses the argument list (arguments already
- * follow the opening '(' which was consumed as part
- * of the function token).
- * -------------------------------------------------------- */
-
-/**
- * @brief Parse a built-in function call
- *
- * The function name token (including opening paren)
- * has already been consumed.  We parse the argument
- * expressions separated by commas and expect a
- * closing ')'.
- */
-static val_t parse_funccall(cli_token *ftok)
-{
-    cli_token *(*cur_func)(void);
-    cli_token *(*advance_func)(void);
-
-    if (parse_error || eval_error) { // Check both error flags
-        return mk_double(0);
-    }
-
-    // Determine which token stream to use based on which error flag is active
-    // This is a bit of a hack, ideally the parser state would be passed explicitly
-    if (parse_mode == 0) { // cli_parse is active
-        cur_func = cur_parse;
-        advance_func = advance_parse;
-    } else { // cli_calc_eval_line is active
-        cur_func = cur_eval;
-        advance_func = advance_eval;
-    }
-
-
-    if (ftok->type == TOK_FUNC_D_D)
-    {
-        /*
-         * Function: double -> double
-         * Also handles image -> image
-         */
-        val_t arg = parse_expr(0);
-        if (parse_error || eval_error)
-        {
-            return mk_double(0);
-        }
-
-        if (cur_func()->type != TOK_RPAREN)
-        {
-            parse_errmsg("Expected ')'");
-            return mk_double(0);
-        }
-        advance_func();
-
-        if (arg.type == VAL_STRING)
-        {
-            /* image -> image via function */
-            if (!check_image(arg.sval))
-            {
-                return mk_string("");
-            }
-            const char *tmpn = alloc_tmpname();
-            arith_image_function_im_im__d_d(
-                arg.sval, tmpn, ftok->fnctptr
-            );
-            if (data.core.Debug > 0)
-            {
-                printf("double_func(double)\n");
-            }
-            return mk_string(tmpn);
-        }
-
-        double darg = to_double(arg);
-        double res  = ((double (*)(double)) ftok->fnctptr)(darg);
-        if (data.core.Debug > 0)
-        {
-            printf("double=func(double)\n");
-        }
-        return mk_double(res);
-    }
-
-    if (ftok->type == TOK_FUNC_DD_D)
-    {
-        /*
-         * Function: (double, double) -> double
-         * Also handles (image, double) -> image
-         * Also handles (image, image) -> image for specific functions
-         */
-        val_t arg1 = parse_expr(0);
-        if (parse_error || eval_error)
-        {
-            return mk_double(0);
-        }
-
-        if (cur_func()->type != TOK_COMMA)
-        {
-            parse_errmsg("Expected ','");
-            return mk_double(0);
-        }
-        advance_func();
-
-        val_t arg2 = parse_expr(0);
-        if (parse_error || eval_error)
-        {
-            return mk_double(0);
-        }
-
-        if (cur_func()->type != TOK_RPAREN)
-        {
-            parse_errmsg("Expected ')'");
-            return mk_double(0);
-        }
-        advance_func();
-
-        if (arg1.type == VAL_STRING && arg2.type == VAL_STRING)
-        {
-            if (!check_image(arg1.sval) || !check_image(arg2.sval))
-            {
-                return mk_string("");
-            }
-            const char *tmpn = alloc_tmpname();
-            if (ftok->fnctptr == fmin)
-            {
-                arith_image_minv(arg1.sval, arg2.sval, tmpn);
-            }
-            else if (ftok->fnctptr == fmax)
-            {
-                arith_image_maxv(arg1.sval, arg2.sval, tmpn);
-            }
-            else if (ftok->fnctptr == fmod)
-            {
-                arith_image_fmod(arg1.sval, arg2.sval, tmpn);
-            }
-            else if (ftok->fnctptr == pow)
-            {
-                arith_image_pow(arg1.sval, arg2.sval, tmpn);
-            }
-            else
-            {
-                parse_errmsg("Unsupported (image, image) function");
-                return mk_string("");
-            }
-            return mk_string(tmpn);
-        }
-
-        if (arg1.type == VAL_STRING)
-        {
-            if (!check_image(arg1.sval))
-            {
-                return mk_string("");
-            }
-            const char *tmpn = alloc_tmpname();
-            arith_image_function_imd_im__dd_d(
-                arg1.sval,
-                to_double(arg2),
-                tmpn,
-                ftok->fnctptr
-            );
-            return mk_string(tmpn);
-        }
-
-        double d1 = to_double(arg1);
-        double d2 = to_double(arg2);
-        return mk_double(((double (*)(double, double)) ftok->fnctptr)(d1, d2));
-    }
-
-    if (ftok->type == TOK_FUNC_DDD_D)
-    {
-        /*
-         * Function: (d, d, d) -> double
-         * Also handles (image, d, d) -> image
-         */
-        val_t arg1 = parse_expr(0);
-        if (parse_error || eval_error)
-        {
-            return mk_double(0);
-        }
-
-        if (cur_func()->type != TOK_COMMA)
-        {
-            parse_errmsg("Expected ','");
-            return mk_double(0);
-        }
-        advance_func();
-
-        val_t arg2 = parse_expr(0);
-        if (parse_error || eval_error)
-        {
-            return mk_double(0);
-        }
-
-        if (cur_func()->type != TOK_COMMA)
-        {
-            parse_errmsg("Expected ','");
-            return mk_double(0);
-        }
-        advance_func();
-
-        val_t arg3 = parse_expr(0);
-        if (parse_error || eval_error)
-        {
-            return mk_double(0);
-        }
-
-        if (cur_func()->type != TOK_RPAREN)
-        {
-            parse_errmsg("Expected ')'");
-            return mk_double(0);
-        }
-        advance_func();
-
-        if (arg1.type == VAL_STRING)
-        {
-            if (!check_image(arg1.sval))
-            {
-                return mk_string("");
-            }
-            const char *tmpn = alloc_tmpname();
-            arith_image_function_imdd_im__ddd_d(
-                arg1.sval,
-                to_double(arg2),
-                to_double(arg3),
-                tmpn,
-                ftok->fnctptr
-            );
-            return mk_string(tmpn);
-        }
-
-        double d1 = to_double(arg1);
-        double d2 = to_double(arg2);
-        double d3 = to_double(arg3);
-        return mk_double(
-            ((double (*)(double, double, double)) ftok->fnctptr)(d1, d2, d3)
-        );
-    }
-
-    if (ftok->type == TOK_FUNC_WHERE)
-    {
-        val_t cond = parse_expr(0);
-        if (parse_error || eval_error)
-            return mk_double(0);
-
-        if (cur_func()->type != TOK_COMMA)
-        {
-            parse_errmsg("Expected ','");
-            return mk_double(0);
-        }
-        advance_func();
-
-        val_t argT = parse_expr(0);
-        if (parse_error || eval_error)
-            return mk_double(0);
-
-        if (cur_func()->type != TOK_COMMA)
-        {
-            parse_errmsg("Expected ','");
-            return mk_double(0);
-        }
-        advance_func();
-
-        val_t argF = parse_expr(0);
-        if (parse_error || eval_error)
-            return mk_double(0);
-
-        if (cur_func()->type != TOK_RPAREN)
-        {
-            parse_errmsg("Expected ')'");
-            return mk_double(0);
-        }
-        advance_func();
-
-        /* If cond is scalar */
-        if (cond.type != VAL_STRING)
-        {
-            double c = to_double(cond);
-            if (c != 0) return argT;
-            else return argF;
-        }
-
-        /* cond is an image. Handle T/F types */
-        char tmp_mask[200], tmp_imask[200], tmp_tpart[200], tmp_fpart[200], tmpn[200];
-        snprintf(tmp_mask, 200, "%s", alloc_tmpname());
-        snprintf(tmp_imask, 200, "%s", alloc_tmpname());
-        snprintf(tmp_tpart, 200, "%s", alloc_tmpname());
-        snprintf(tmp_fpart, 200, "%s", alloc_tmpname());
-        snprintf(tmpn, 200, "%s", alloc_tmpname());
-
-        if (!check_image(cond.sval))
-            return mk_string("");
-
-        /* mask = (cond != 0) */
-        arith_image_csttestne(cond.sval, 0.0, tmp_mask);
-        /* imask = (cond == 0) */
-        arith_image_cstteste(cond.sval, 0.0, tmp_imask);
-
-        if (argT.type == VAL_STRING)
-        {
-            if (!check_image(argT.sval)) return mk_string("");
-            arith_image_mult(argT.sval, tmp_mask, tmp_tpart);
-        }
-        else
-        {
-            arith_image_cstmult(tmp_mask, to_double(argT), tmp_tpart);
-        }
-
-        if (argF.type == VAL_STRING)
-        {
-            if (!check_image(argF.sval)) return mk_string("");
-            arith_image_mult(argF.sval, tmp_imask, tmp_fpart);
-        }
-        else
-        {
-            arith_image_cstmult(tmp_imask, to_double(argF), tmp_fpart);
-        }
-
-        arith_image_add(tmp_tpart, tmp_fpart, tmpn);
-        return mk_string(tmpn);
-    }
-
-    if (ftok->type == TOK_FUNC_IMIM_D)
-    {
-        val_t arg1 = parse_expr(0);
-        if (parse_error || eval_error) return mk_double(0);
-        if (cur_func()->type != TOK_COMMA) { parse_errmsg("Expected ','"); return mk_double(0); }
-        advance_func();
-        val_t arg2 = parse_expr(0);
-        if (parse_error || eval_error) return mk_double(0);
-        if (cur_func()->type != TOK_RPAREN) { parse_errmsg("Expected ')'"); return mk_double(0); }
-        advance_func();
-
-        if (arg1.type != VAL_STRING || arg2.type != VAL_STRING) {
-            parse_errmsg("Expected two image arguments");
-            return mk_double(0);
-        }
-        if (!check_image(arg1.sval) || !check_image(arg2.sval)) return mk_double(0);
-
-        // Call ftok->fnctptr for vector operation (e.g. dot())
-        double res = ((double (*)(const char *, const char *)) ftok->fnctptr)(arg1.sval, arg2.sval);
-        return mk_double(res);
-    }
-
-    if (ftok->type == TOK_FUNC_IM_D)
-    {
-        /* Function: image -> double */
-        val_t arg = parse_expr(0);
-        if (parse_error || eval_error)
-        {
-            return mk_double(0);
-        }
-
-        if (cur_func()->type != TOK_RPAREN)
-        {
-            parse_errmsg("Expected ')'");
-            return mk_double(0);
-        }
-        advance_func();
-
-        if (arg.type != VAL_STRING)
-        {
-            parse_errmsg(
-                "Expected image argument"
-            );
-            return mk_double(0);
-        }
-        if (!check_image(arg.sval))
-        {
-            return mk_double(0);
-        }
-
-        double res = ((double (*)(const char *)) ftok->fnctptr)(arg.sval);
-        if (data.core.Debug > 0)
-        {
-            printf("double=func(image)\n");
-        }
-        return mk_double(res);
-    }
-
-    if (ftok->type == TOK_FUNC_IMD_D)
-    {
-        /* Function: (image, double) -> double */
-        val_t arg1 = parse_expr(0);
-        if (parse_error || eval_error)
-        {
-            return mk_double(0);
-        }
-
-        if (cur_func()->type != TOK_COMMA)
-        {
-            parse_errmsg("Expected ','");
-            return mk_double(0);
-        }
-        advance_func();
-
-        val_t arg2 = parse_expr(0);
-        if (parse_error || eval_error)
-        {
-            return mk_double(0);
-        }
-
-        if (cur_func()->type != TOK_RPAREN)
-        {
-            parse_errmsg("Expected ')'");
-            return mk_double(0);
-        }
-        advance_func();
-
-        if (arg1.type != VAL_STRING)
-        {
-            parse_errmsg(
-                "Expected image argument"
-            );
-            return mk_double(0);
-        }
-        if (!check_image(arg1.sval))
-        {
-            return mk_double(0);
-        }
-
-        double res = ((double (*)(const char *, double)) ftok->fnctptr)(
-            arg1.sval, to_double(arg2)
-        );
-        return mk_double(res);
-    }
-
-    /* strlen(string) -> long */
-    if (ftok->type == TOK_FUNC_S_D)
-    {
-        val_t arg = parse_expr(0);
-        if (parse_error || eval_error)
-        {
-            return mk_double(0);
-        }
-        if (cur_func()->type != TOK_RPAREN)
-        {
-            parse_errmsg("Expected ')'");
-            return mk_double(0);
-        }
-        advance_func();
-
-        /* Argument is a variable name or
-         * image name stored in sval */
-        if (arg.type == VAL_STRING)
-        {
-            /* Try CLI var first */
-            const char *cv =
-                cli_var_get(arg.sval);
-            if (cv != NULL)
-            {
-                return mk_long(
-                    (long) strlen(cv)
-                );
-            }
-            /* Fallback: length of name */
-            return mk_long(
-                (long) strlen(arg.sval)
-            );
-        }
-        /* If numeric, convert to string */
-        char numstr[64];
-        if (arg.type == VAL_LONG)
-        {
-            snprintf(numstr, 64,
-                     "%ld", arg.lval);
-        }
-        else
-        {
-            snprintf(numstr, 64,
-                     "%g", arg.dval);
-        }
-        return mk_long(
-            (long) strlen(numstr)
-        );
-    }
-
-    /* toupper(s), tolower(s) -> string */
-    if (ftok->type == TOK_FUNC_S_S)
-    {
-        val_t arg = parse_expr(0);
-        if (parse_error || eval_error)
-        {
-            return mk_double(0);
-        }
-        if (cur_func()->type != TOK_RPAREN)
-        {
-            parse_errmsg("Expected ')'");
-            return mk_double(0);
-        }
-        advance_func();
-
-        /* Resolve string value */
-        const char *sv = NULL;
-        char numbuf[64];
-        if (arg.type == VAL_STRING)
-        {
-            const char *cv =
-                cli_var_get(arg.sval);
-            sv = cv ? cv : arg.sval;
-        }
-        else if (arg.type == VAL_LONG)
-        {
-            snprintf(numbuf, 64,
-                     "%ld", arg.lval);
-            sv = numbuf;
-        }
-        else
-        {
-            snprintf(numbuf, 64,
-                     "%g", arg.dval);
-            sv = numbuf;
-        }
-
-        char result[CLI_CALC_TOKEN_MAXLEN];
-        strncpy(result, sv,
-                sizeof(result) - 1);
-        result[sizeof(result) - 1] = '\0';
-
-        const char *fn = ftok->sval;
-        if (strncmp(fn, "toupper(", 8) == 0)
-        {
-            for (int i = 0;
-                 result[i]; i++)
-            {
-                result[i] = (char) toupper(
-                    (unsigned char) result[i]);
-            }
-        }
-        else /* tolower */
-        {
-            for (int i = 0;
-                 result[i]; i++)
-            {
-                result[i] = (char) tolower(
-                    (unsigned char) result[i]);
-            }
-        }
-
-        /* Store as temp CLI var */
-        const char *tmpn = alloc_tmpname();
-        cli_var_set(tmpn, result);
-        return mk_string(tmpn);
-    }
-
-    /* substr(s, off, len) -> string */
-    if (ftok->type == TOK_FUNC_SDD_S)
-    {
-        val_t arg1 = parse_expr(0);
-        if (parse_error || eval_error)
-        {
-            return mk_double(0);
-        }
-        if (cur_func()->type != TOK_COMMA)
-        {
-            parse_errmsg("substr: need 3 args");
-            return mk_double(0);
-        }
-        advance_func();
-        val_t arg2 = parse_expr(0);
-        if (parse_error || eval_error)
-        {
-            return mk_double(0);
-        }
-        if (cur_func()->type != TOK_COMMA)
-        {
-            parse_errmsg("substr: need 3 args");
-            return mk_double(0);
-        }
-        advance_func();
-        val_t arg3 = parse_expr(0);
-        if (parse_error || eval_error)
-        {
-            return mk_double(0);
-        }
-        if (cur_func()->type != TOK_RPAREN)
-        {
-            parse_errmsg("Expected ')'");
-            return mk_double(0);
-        }
-        advance_func();
-
-        const char *sv = NULL;
-        if (arg1.type == VAL_STRING)
-        {
-            const char *cv =
-                cli_var_get(arg1.sval);
-            sv = cv ? cv : arg1.sval;
-        }
-        else
-        {
-            parse_errmsg(
-                "substr: first arg "
-                "must be string");
-            return mk_double(0);
-        }
-
-        int slen = (int) strlen(sv);
-        int off = (int) to_double(arg2);
-        int len = (int) to_double(arg3);
-        if (off < 0) { off = 0; }
-        if (off > slen) { off = slen; }
-        if (len < 0) { len = 0; }
-        if (off + len > slen)
-        {
-            len = slen - off;
-        }
-
-        char result[CLI_CALC_TOKEN_MAXLEN];
-        memcpy(result, sv + off,
-               (size_t) len);
-        result[len] = '\0';
-
-        const char *tmpn = alloc_tmpname();
-        cli_var_set(tmpn, result);
-        return mk_string(tmpn);
-    }
-
-    /* replace(s, old, new) -> string */
-    if (ftok->type == TOK_FUNC_SSS_S)
-    {
-        val_t arg1 = parse_expr(0);
-        if (parse_error || eval_error)
-        {
-            return mk_double(0);
-        }
-        if (cur_func()->type != TOK_COMMA)
-        {
-            parse_errmsg(
-                "replace: need 3 args");
-            return mk_double(0);
-        }
-        advance_func();
-        val_t arg2 = parse_expr(0);
-        if (parse_error || eval_error)
-        {
-            return mk_double(0);
-        }
-        if (cur_func()->type != TOK_COMMA)
-        {
-            parse_errmsg(
-                "replace: need 3 args");
-            return mk_double(0);
-        }
-        advance_func();
-        val_t arg3 = parse_expr(0);
-        if (parse_error || eval_error)
-        {
-            return mk_double(0);
-        }
-        if (cur_func()->type != TOK_RPAREN)
-        {
-            parse_errmsg("Expected ')'");
-            return mk_double(0);
-        }
-        advance_func();
-
-        const char *s1 = NULL;
-        const char *s2 = NULL;
-        const char *s3 = NULL;
-        if (arg1.type == VAL_STRING)
-        {
-            const char *cv =
-                cli_var_get(arg1.sval);
-            s1 = cv ? cv : arg1.sval;
-        }
-        if (arg2.type == VAL_STRING)
-        {
-            const char *cv =
-                cli_var_get(arg2.sval);
-            s2 = cv ? cv : arg2.sval;
-        }
-        if (arg3.type == VAL_STRING)
-        {
-            const char *cv =
-                cli_var_get(arg3.sval);
-            s3 = cv ? cv : arg3.sval;
-        }
-        if (!s1 || !s2 || !s3)
-        {
-            parse_errmsg(
-                "replace: all args "
-                "must be strings");
-            return mk_double(0);
-        }
-
-        char result[CLI_CALC_TOKEN_MAXLEN];
-        result[0] = '\0';
-        int s2len = (int) strlen(s2);
-        if (s2len == 0)
-        {
-            strncpy(result, s1,
-                    sizeof(result) - 1);
-            result[sizeof(result) - 1]
-                = '\0';
-        }
-        else
-        {
-            const char *p = s1;
-            char *w = result;
-            char *end = result
-                + sizeof(result) - 1;
-            while (*p && w < end)
-            {
-                if (strncmp(p, s2,
-                    (size_t) s2len) == 0)
-                {
-                    int rlen =
-                        (int) strlen(s3);
-                    if (w + rlen < end)
-                    {
-                        memcpy(w, s3,
-                            (size_t) rlen);
-                        w += rlen;
-                    }
-                    p += s2len;
-                }
-                else
-                {
-                    *w++ = *p++;
-                }
-            }
-            *w = '\0';
-        }
-
-        const char *tmpn = alloc_tmpname();
-        cli_var_set(tmpn, result);
-        return mk_string(tmpn);
-    }
-
-    /* hex(n), oct(n), bin(n) -> string */
-    if (ftok->type == TOK_FUNC_D_S)
-    {
-        val_t arg = parse_expr(0);
-        if (parse_error || eval_error)
-        {
-            return mk_double(0);
-        }
-        if (cur_func()->type != TOK_RPAREN)
-        {
-            parse_errmsg("Expected ')'");
-            return mk_double(0);
-        }
-        advance_func();
-
-        long iv = (arg.type == VAL_LONG)
-            ? arg.lval
-            : (long) to_double(arg);
-
-        char result[CLI_CALC_TOKEN_MAXLEN];
-        const char *fn = ftok->sval;
-        if (strncmp(fn, "hex(", 4) == 0)
-        {
-            snprintf(result,
-                     sizeof(result),
-                     "0x%lx", iv);
-        }
-        else if (strncmp(fn, "oct(", 4)
-                 == 0)
-        {
-            snprintf(result,
-                     sizeof(result),
-                     "0o%lo", iv);
-        }
-        else /* bin */
-        {
-            char *w = result;
-            *w++ = '0';
-            *w++ = 'b';
-            if (iv == 0)
-            {
-                *w++ = '0';
-            }
-            else
-            {
-                long v = iv;
-                char bits[65];
-                int bi = 0;
-                while (v > 0 && bi < 64)
-                {
-                    bits[bi++] =
-                        (v & 1) ? '1' : '0';
-                    v >>= 1;
-                }
-                for (int i = bi - 1;
-                     i >= 0; i--)
-                {
-                    *w++ = bits[i];
-                }
-            }
-            *w = '\0';
-        }
-
-        const char *tmpn = alloc_tmpname();
-        cli_var_set(tmpn, result);
-        return mk_string(tmpn);
-    }
-
-    parse_errmsg("Unknown function type");
-    return mk_double(0);
-}
-
-/* --------------------------------------------------------
- * Primary expression parser
- *
- * Handles atoms: numbers, parenthesized expressions,
- * unary minus, function calls, and identifiers
- * (with optional assignment).
- * -------------------------------------------------------- */
-
-/**
- * @brief Parse a primary (non-operator) expression
- */
-static val_t parse_primary(void)
-{
-    cli_token *t;
-    cli_token *(*cur_func)(void);
-    cli_token *(*advance_func)(void);
-
-    if (parse_error || eval_error) { // Check both error flags
-        return mk_long(0);
-    }
-
-    // Determine which token stream to use based on which error flag is active
-    if (parse_mode == 0) { // cli_parse is active
-        cur_func = cur_parse;
-        advance_func = advance_parse;
-    } else { // cli_calc_eval_line is active
-        cur_func = cur_eval;
-        advance_func = advance_eval;
-    }
-
-    t = cur_func();
-
-    /* number literals */
-    if (t->type == TOK_LONG)
-    {
-        advance_func();
-        if (data.core.Debug > 0)
-        {
-            printf("this is a long\n");
-        }
-        return mk_long(t->val_l);
-    }
-
-    if (t->type == TOK_DOUBLE)
-    {
-        advance_func();
-        if (data.core.Debug > 0)
-        {
-            printf("this is a double\n");
-        }
-        return mk_double(t->val_d);
-    }
-
-    /* unary minus */
-    if (t->type == TOK_OP_MINUS)
-    {
-        advance_func();
-        val_t v = parse_primary();
-        if (parse_error || eval_error)
-        {
-            return mk_long(0);
-        }
-        if (v.type == VAL_LONG)
-        {
-            if (data.core.Debug > 0)
-            {
-                printf("-long\n");
-            }
-            return mk_long(-v.lval);
-        }
-        if (v.type == VAL_DOUBLE)
-        {
-            if (data.core.Debug > 0)
-            {
-                printf("-double\n");
-            }
-            return mk_double(-v.dval);
-        }
-        if (v.type == VAL_STRING)
-        {
-            if (data.core.Debug > 0)
-            {
-                printf("-image\n");
-            }
-            if (!check_image(v.sval))
-            {
-                return mk_double(0);
-            }
-            const char *tmpn = alloc_tmpname();
-            arith_image_cstsubm(v.sval, 0.0, tmpn);
-            return mk_string(tmpn);
-        }
-        parse_errmsg(
-            "Cannot negate expression"
-        );
-        return mk_double(0);
-    }
-
-    /* unary bitwise NOT */
-    if (t->type == TOK_OP_BNOT)
-    {
-        advance_func();
-        val_t v = parse_primary();
-        if (parse_error || eval_error)
-        {
-            return mk_long(0);
-        }
-        long iv = (v.type == VAL_LONG)
-            ? v.lval : (long) v.dval;
-        return mk_long(~iv);
-    }
-
-    /* parenthesized expression */
-    if (t->type == TOK_LPAREN)
-    {
-        advance_func();
-        val_t v = parse_expr(0);
-        if (parse_error || eval_error)
-        {
-            return v;
-        }
-        if (cur_func()->type != TOK_RPAREN)
-        {
-            parse_errmsg("Expected ')'");
-            return mk_long(0);
-        }
-        advance_func();
-        return v;
-    }
-
-    /* function calls */
-    if (t->type == TOK_FUNC_D_D
-        || t->type == TOK_FUNC_DD_D
-        || t->type == TOK_FUNC_DDD_D
-        || t->type == TOK_FUNC_IM_D
-        || t->type == TOK_FUNC_IMD_D
-        || t->type == TOK_FUNC_WHERE
-        || t->type == TOK_FUNC_IMIM_D
-        || t->type == TOK_FUNC_S_D
-        || t->type == TOK_FUNC_S_S
-        || t->type == TOK_FUNC_SDD_S
-        || t->type == TOK_FUNC_SSS_S
-        || t->type == TOK_FUNC_D_S)
-    {
-        advance_func();
-        return parse_funccall(t);
-    }
-
-    /* existing variable: might be assignment */
-    if (t->type == TOK_VAR)
-    {
-        advance_func();
-        cli_token_type aop = cur_func()->type;
-        if (aop == TOK_EQUAL
-            || aop == TOK_OP_PLUS_EQ
-            || aop == TOK_OP_MINUS_EQ
-            || aop == TOK_OP_STAR_EQ
-            || aop == TOK_OP_SLASH_EQ)
-        {
-            advance_func();
-            val_t v = parse_expr(0);
-            if (parse_error || eval_error)
-            {
-                return v;
-            }
-
-            /* compound: fetch old, apply op */
-            if (aop != TOK_EQUAL)
-            {
-                long vid = variable_ID(t->sval);
-                if (vid == -1)
-                {
-                    parse_errmsg(
-                        "Variable not found"
-                        " for compound assign"
-                    );
-                    return mk_double(0);
-                }
-                double old;
-                if (data.core.variable[vid]
-                    .type == 1)
-                {
-                    old = (double)
-                        data.core.variable[vid]
-                            .value.l;
-                }
-                else
-                {
-                    old = data.core.variable[vid]
-                        .value.f;
-                }
-                double nv = to_double(v);
-                switch (aop)
-                {
-                    case TOK_OP_PLUS_EQ:
-                        nv = old + nv; break;
-                    case TOK_OP_MINUS_EQ:
-                        nv = old - nv; break;
-                    case TOK_OP_STAR_EQ:
-                        nv = old * nv; break;
-                    case TOK_OP_SLASH_EQ:
-                        if (nv == 0.0)
-                        {
-                            parse_errmsg(
-                                "Division by zero"
-                            );
-                            return mk_double(0);
-                        }
-                        nv = old / nv;
-                        break;
-                    default:
-                        break;
-                }
-                /* check if result is integer */
-                if (v.type == VAL_LONG
-                    && data.core.variable[vid]
-                        .type == 1
-                    && aop != TOK_OP_SLASH_EQ)
-                {
-                    v = mk_long((long) nv);
-                }
-                else
-                {
-                    v = mk_double(nv);
-                }
-            }
-
-            if (v.type == VAL_STRING)
-            {
-                /* var = image -> rename */
-                chname_image_ID(
-                    v.sval, t->sval
-                );
-                if (data.core.Debug > 0)
-                {
-                    printf("changing name\n");
-                }
-                return mk_string(t->sval);
-            }
-            if (v.type == VAL_LONG)
-            {
-                create_variable_long_ID(
-                    t->sval, v.lval
-                );
-                return mk_long(v.lval);
-            }
-            create_variable_ID(
-                t->sval, to_double(v)
-            );
-            return mk_double(to_double(v));
-        }
-        /* just a variable reference */
-        long vID = variable_ID(t->sval);
-        if (vID == -1)
-        {
-            char msg[2048];
-            snprintf(msg, sizeof(msg),
-                     "Variable '%s' not found",
-                     t->sval);
-            parse_errmsg(msg);
-            return mk_double(0);
-        }
-        if (data.core.variable[vID].type == 1)
-        {
-            return mk_long(
-                data.core.variable[vID].value.l
-            );
-        }
-        return mk_double(
-            data.core.variable[vID].value.f
-        );
-    }
-
-    /* new variable: must be assignment, or it's
-     * a new string/image name */
-    if (t->type == TOK_NVAR)
-    {
-        advance_func();
-        if (cur_func()->type == TOK_EQUAL)
-        {
-            advance_func();
-            val_t v = parse_expr(0);
-            if (parse_error || eval_error)
-            {
-                return v;
-            }
-            if (v.type == VAL_STRING)
-            {
-                if (image_ID(v.sval, data.core.image, data.core.NB_MAX_IMAGE) == -1) {
-                    parse_errmsg("Source image does not exist");
-                    return mk_double(0);
-                }
-                chname_image_ID(
-                    v.sval, t->sval
-                );
-                if (data.core.Debug > 0)
-                {
-                    printf("changing name\n");
-                }
-                return mk_string(t->sval);
-            }
-            if (v.type == VAL_LONG)
-            {
-                if (data.core.Debug > 0)
-                {
-                    printf("creating long\n");
-                }
-            }
-            else
-            {
-                if (data.core.Debug > 0)
-                {
-                    printf("creating double\n");
-                }
-            }
-            if (v.type == VAL_LONG)
-            {
-                create_variable_long_ID(
-                    t->sval, v.lval
-                );
-                char numv[64];
-                snprintf(
-                    numv, 64, "%ld", v.lval
-                );
-                if (parse_mode == 1)
-                {
-                    cli_var_set(t->sval, numv);
-                }
-                return mk_long(v.lval);
-            }
-            create_variable_ID(
-                t->sval, to_double(v)
-            );
-            {
-                char numv[64];
-                snprintf(
-                    numv, 64, "%g",
-                    to_double(v)
-                );
-                if (parse_mode == 1)
-                {
-                    cli_var_set(t->sval, numv);
-                }
-            }
-            return mk_double(to_double(v));
-        }
-        /* standalone new variable/image name */
-        // This path is only for cli_parse, not cli_calc_eval_line
-        if (parse_mode == 0) { // If not in eval_line context
-            data.cmdargtoken[data.cmdNBarg].type =
-                CMDARGTOKEN_TYPE_STRING;
-            if (data.core.Debug > 0)
-            {
-                printf(
-                    "this is a string "
-                    "(new variable/image)\n"
-                );
-            }
-        }
-        return mk_string(t->sval);
-    }
-
-    /* existing image: might be assignment */
-    if (t->type == TOK_IMAGE)
-    {
-        advance_func();
-        if (cur_func()->type == TOK_EQUAL)
-        {
-            advance_func();
-            val_t v = parse_expr(0);
-            if (parse_error || eval_error)
-            {
-                return v;
-            }
-            if (v.type == VAL_STRING)
-            {
-                if (image_ID(v.sval, data.core.image, data.core.NB_MAX_IMAGE) == -1) {
-                    parse_errmsg("Source image does not exist");
-                    return mk_double(0);
-                }
-                delete_image_ID(
-                    t->sval,
-                    1
-                );
-                chname_image_ID(
-                    v.sval, t->sval
-                );
-                if (data.core.Debug > 0)
-                {
-                    printf("changing name\n");
-                }
-                return mk_string(t->sval);
-            }
-            parse_errmsg(
-                "Cannot assign scalar to image"
-            );
-            return mk_double(0);
-        }
-        // This path is only for cli_parse, not cli_calc_eval_line
-        if (!eval_error) { // If not in eval_line context
-            data.cmdargtoken[data.cmdNBarg].type =
-                CMDARGTOKEN_TYPE_EXISTINGIMAGE;
-            if (data.core.Debug > 0)
-            {
-                printf(
-                    "this is a string "
-                    "(existing image)\n"
-                );
-            }
-        }
-        /* Check for slice bracket syntax.
-         * Materialize the sliced region into
-         * a temporary image and return that
-         * name instead. */
-        if (strchr(t->sval, '[') != NULL)
-        {
-            char bare[200];
-            char btext[200];
-            int has_brk =
-                imgid_slice_split_name(
-                    t->sval,
-                    bare, (int) sizeof(bare),
-                    btext,
-                    (int) sizeof(btext));
-            if (has_brk)
-            {
-                imageID srcid = image_ID(
-                    bare,
-                    data.core.image,
-                    data.core.NB_MAX_IMAGE);
-                if (srcid == -1)
-                {
-                    parse_errmsg(
-                        "Source image "
-                        "not found");
-                    return mk_double(0);
-                }
-                IMAGE *srcim =
-                    &data.core.image[srcid];
-                IMGID_SLICE slc =
-                    imgid_slice_parse(btext);
-                if (slc.error)
-                {
-                    parse_errmsg(
-                        slc.errmsg);
-                    return mk_double(0);
-                }
-                uint32_t outsz[3] = {0};
-                int snax =
-                    srcim->md[0].naxis;
-                uint32_t ssz[3];
-                for (int a = 0;
-                     a < snax && a < 3;
-                     a++)
-                {
-                    ssz[a] =
-                        srcim->md[0].size[a];
-                }
-                if (imgid_slice_output_size(
-                        &slc, snax,
-                        ssz, outsz) != 0)
-                {
-                    parse_errmsg(
-                        "Bad slice dims");
-                    return mk_double(0);
-                }
-                /* Count output axes */
-                int onax = 0;
-                for (int a = 0; a < 3; a++)
-                {
-                    if (outsz[a] > 0)
-                    {
-                        onax = a + 1;
-                    }
-                }
-                if (onax == 0)
-                {
-                    onax = 1;
-                }
-                const char *tmpn =
-                    alloc_tmpname();
-                imageID tid =
-                    create_image_ID(
-                        tmpn,
-                        onax,
-                        outsz,
-                        srcim->md[0].datatype,
-                        0, 10, 0, NULL);
-                if (tid != -1)
-                {
-                    IMGID simg;
-                    memset(&simg, 0,
-                           sizeof(simg));
-                    simg.ID = srcid;
-                    simg.im = srcim;
-                    simg.md = srcim->md;
-                    simg.slice = slc;
-                    simg.slice_im =
-                        &data.core.image[tid];
-                    imgid_slice_materialize(
-                        &simg);
-                }
-                return mk_string(tmpn);
-            }
-        }
-        return mk_string(t->sval);
-    }
-
-    /* command */
-    if (t->type == TOK_COMMAND)
-    {
-        advance_func();
-        // This path is only for cli_parse, not cli_calc_eval_line
-        if (parse_mode == 0) { // If not in eval_line context
-            data.cmdargtoken[data.cmdNBarg].type =
-                CMDARGTOKEN_TYPE_COMMAND;
-            if (data.core.Debug > 0)
-            {
-                printf(
-                    "this is a string (command)\n"
-                );
-            }
-        }
-        return mk_string(t->sval);
-    }
-
-    parse_errmsg("Unexpected token");
-    return mk_long(0);
-}
-
-/* --------------------------------------------------------
- * Expression parser (precedence climbing)
- * -------------------------------------------------------- */
-
-/**
- * @brief Parse an expression with minimum precedence
- *
- * Implements Pratt / precedence-climbing:
- *  1. Parse a primary (atom)
- *  2. While the next token is a binary operator with
- *     precedence >= min_prec, consume it and parse
- *     the right-hand side with appropriate precedence
- */
-static val_t parse_expr(int min_prec)
-{
-    cli_token *(*cur_func)(void);
-    cli_token *(*advance_func)(void);
-
-    if (parse_error || eval_error) { // Check both error flags
-        return mk_long(0);
-    }
-
-    // Determine which token stream to use based on which error flag is active
-    if (parse_mode == 0) { // cli_parse is active
-        cur_func = cur_parse;
-        advance_func = advance_parse;
-    } else { // cli_calc_eval_line is active
-        cur_func = cur_eval;
-        advance_func = advance_eval;
-    }
-
-
-    val_t left = parse_primary();
-    if (parse_error || eval_error)
-    {
-        return left;
-    }
-
-    for (;;)
-    {
-        cli_token_type op = cur_func()->type;
-        int prec = get_prec(op);
-
-        if (prec < min_prec)
-        {
-            break;
-        }
-
-        advance_func();
-
-        /* Ternary ? : operator */
-        if (op == TOK_OP_QUESTION)
-        {
-            val_t true_val = parse_expr(-1);
-            if (parse_error || eval_error)
-            {
-                return left;
-            }
-            if (cur_func()->type
-                != TOK_OP_COLON)
-            {
-                parse_errmsg(
-                    "Expected ':' in ternary"
-                );
-                return mk_double(0);
-            }
-            advance_func();
-            val_t false_val = parse_expr(-1);
-            if (parse_error || eval_error)
-            {
-                return left;
-            }
-            /* Scalar ternary */
-            if (left.type != VAL_STRING)
-            {
-                double c = to_double(left);
-                left = (c != 0.0)
-                    ? true_val : false_val;
-            }
-            else
-            {
-                /* Image ternary -> where() */
-                if (!check_image(left.sval))
-                {
-                    return mk_string("");
-                }
-                const char *tmask =
-                    alloc_tmpname();
-                const char *timask =
-                    alloc_tmpname();
-                const char *tpart =
-                    alloc_tmpname();
-                const char *fpart =
-                    alloc_tmpname();
-                const char *tmpn =
-                    alloc_tmpname();
-                arith_image_csttestne(
-                    left.sval, 0.0,
-                    tmask
-                );
-                arith_image_cstteste(
-                    left.sval, 0.0,
-                    timask
-                );
-                if (true_val.type
-                    == VAL_STRING)
-                {
-                    if (!check_image(
-                            true_val.sval))
-                    {
-                        return mk_string("");
-                    }
-                    arith_image_mult(
-                        true_val.sval,
-                        tmask, tpart
-                    );
-                }
-                else
-                {
-                    arith_image_cstmult(
-                        tmask,
-                        to_double(true_val),
-                        tpart
-                    );
-                }
-                if (false_val.type
-                    == VAL_STRING)
-                {
-                    if (!check_image(
-                            false_val.sval))
-                    {
-                        return mk_string("");
-                    }
-                    arith_image_mult(
-                        false_val.sval,
-                        timask, fpart
-                    );
-                }
-                else
-                {
-                    arith_image_cstmult(
-                        timask,
-                        to_double(false_val),
-                        fpart
-                    );
-                }
-                arith_image_add(
-                    tpart, fpart, tmpn
-                );
-                left = mk_string(tmpn);
-            }
-            continue;
-        }
-
-        int next_min = is_right_assoc(op)
-                       ? prec
-                       : prec + 1;
-        val_t right = parse_expr(next_min);
-        if (parse_error || eval_error)
-        {
-            return left;
-        }
-
-        left = eval_binop(op, left, right);
-        if (parse_error || eval_error)
-        {
-            return left;
-        }
-    }
-
-    return left;
-}
 
 /* --------------------------------------------------------
  * Public entry point
